refactor: move things to utils and other stuff
